requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
altair==5.0.1
ansicon==1.89.0
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
arrow==1.2.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
av==10.0.0
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cmaes==0.9.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
contourpy==1.1.0
croniter==1.3.15
ctranslate2==3.16.0
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.0
docopt==0.6.2
einops==0.6.1
exceptiongroup==1.1.1
fastapi==0.88.0
faster-whisper==0.6.0
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
future==0.18.3
gradio==3.23.0
gradio_client==0.2.7
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
humanfriendly==10.0
HyperPyYAML==1.2.1
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.2.0
jsonschema==4.17.3
julius==0.2.7
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
lightning==2.0.3
lightning-cloud==0.5.36
lightning-utilities==0.8.0
linkify-it-py==2.0.2
llvmlite==0.40.1rc1
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
networkx==3.0
nltk==3.8.1
numba==0.57.0
numpy==1.24.1
omegaconf==2.3.0
onnxruntime==1.15.1
optuna==3.2.0
ordered-set==4.1.0
orjson==3.9.1
packaging==23.1
pandas==2.0.2
Pillow==9.3.0
pipreqs==0.4.13
pooch==1.6.0
primePy==1.3
protobuf==4.23.3
psutil==5.9.5
pyannote.audio==2.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.0
pyreadline3==3.4.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.3
pytorch-metric-learning==2.2.0
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
regex==2023.6.3
requests==2.28.1
rich==13.4.2
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
semver==3.0.1
sentencepiece==0.1.99
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
speechbrain==0.5.14
SQLAlchemy==2.0.16
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tabulate==0.9.0
tensorboardX==2.6.1
threadpoolctl==3.1.0
tokenizers==0.13.3
toolz==0.12.0
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchmetrics==0.11.4
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typer==0.9.0
typing_extensions==4.4.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.13
uvicorn==0.22.0
wcwidth==0.2.6
websocket-client==1.6.0
websockets==11.0.3
whisperx==3.1.1
yarg==0.1.9
yarl==1.9.2